inferno::flamegraph

Module color

Color palettes and options for flame graph generation.

Structs

PaletteMap

Mapping of the association between a function name and the color used when drawing information from this function.

SearchColor

SearchColor::default() is rgb(230,0,230).

Enums

BackgroundColor

A flame graph background color.

BasicPalette

A plain color palette in which the color is not chosen based on function semantics.

MultiPalette

A semantic color palette in which different hues are used to signify semantic aspects of different function names (kernel functions, JIT functions, etc.).

Palette

A flame graph color palette.

StrokeColor

StrokeColor::default() is None.

Functions

parse_hex_color

Parses a string as a hex color, returning None if it is an invalid color

Type Aliases

Color

A re-export of RGB8 from the rgb crate.