inferno

Module collapse

Stack collapsing for various input formats.

See the crate-level documentation for details.

Modules

dtrace

Stack collapsing for the output of dtrace.

ghcprof

Stack collapsing for the output of the GHC’s built-in profiler.

guess

Attempts to use whichever Collapse implementation is appropriate for a given input

perf

Stack collapsing for the output of perf script.

recursive

Collapse direct recursive backtraces.

sample

Stack collapsing for the output of sample on macOS.

vsprof

Stack collapsing for the output of the Visual Studio built-in profiler.

vtune

Stack collapsing for the output of VTune.

xctrace

Stack collapsing for the output of the xctrace.

Traits

Collapse

The abstract behavior of stack collapsing.