Module inferno::collapse

Stack collapsing for various input formats.

See the crate-level documentation for details.

Modules

dtrace
Stack collapsing for the output of dtrace.

guess
Attempts to use whichever Collapse implementation is appropriate for a given input

perf
Stack collapsing for the output of perf script.

sample
Stack collapsing for the output of sample on macOS.

vtune
Stack collapsing for the output of VTune.

Traits

Collapse
The abstract behavior of stack collapsing.