inferno::differential

Function from_readers

pub fn from_readers<R1, R2, W>(
    opt: Options,
    before: R1,
    after: R2,
    writer: W,
) -> Result<()>
where
    R1: BufRead,
    R2: BufRead,
    W: Write,

Produce an output that can be used to generate a differential flame graph.

The readers are expected to contain folded stack lines of before and after profiles with the following whitespace-separated fields:

A semicolon-separated list of frame names (e.g., main;foo;bar;baz).

A sample count for the given stack.

The output written to the writer will be similar to the inputs, except there will be two sample count columns – one for each profile.