inferno

Module flamegraph

Tools for producing flame graphs from folded stack traces.

See the crate-level documentation for details.

Re-exports

pub use self::color::Palette;

Modules

color

Color palettes and options for flame graph generation.

defaults

Default values for Options.

Structs

FuncFrameAttrsMap

Provides a way to customize the attributes on the SVG elements for a frame.

Options

Configure the flame graph.

Enums

Direction

The direction the plot should grow.

TextTruncateDirection

The direction text is truncated when it’s too long.

Functions

from_files

Produce a flame graph from files that contain folded stack lines and write the result to provided writer.

from_lines

Produce a flame graph from an iterator over folded stack lines.

from_reader

Produce a flame graph from a reader that contains a sequence of folded stack lines.

from_readers

Produce a flame graph from a set of readers that contain folded stack lines.